erp-backend/requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ─── Database ─────────────────────────────────────────────────────────────────
asyncpg==0.30.0

# ─── Validation & Settings ────────────────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
email-validator==2.2.0

# ─── Security & Auth ──────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.1

# ─── HTTP Client (tests & internal) ──────────────────────────────────────────
httpx==0.28.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.2
